refactor: Use `LemonTabs` rather than `LemonSegmentedSelect`
Makes it look more like the rest of the UI
