+++ frontend/src/queries/nodes/WebVitals/WebVitals.tsx
@@ -89,7 +89,7 @@ export function WebVitals(props: {
                 />
             </div>
 
-            <div className="flex flex-row gap-2 p-4">
+            <div className="flex flex-col sm:flex-row gap-2 p-4">
                 <WebVitalsContent webVitalsQueryResponse={webVitalsQueryResponse} />
                 <div className="flex-1">
                     <Query query={webVitalsMetricQuery} readOnly embedded />
