+++ frontend/src/scenes/web-analytics/WebAnalyticsScene.tsx
@@ -1,8 +1,8 @@
 import { PageHeader } from 'lib/components/PageHeader'
 import { SceneExport } from 'scenes/sceneTypes'
+import { WebAnalyticsDashboard } from 'scenes/web-analytics/WebAnalyticsDashboard'
 import { webAnalyticsLogic } from 'scenes/web-analytics/webAnalyticsLogic'
 import { WebAnalyticsMenu } from 'scenes/web-analytics/WebAnalyticsMenu'
-import { WebAnalyticsDashboard } from 'scenes/web-analytics/WebDashboard'
 
 export function WebAnalyticsScene(): JSX.Element {
     return (
